Add AbstractConfig
