+++ src/Dotmailer/Config.php
@@ -4,22 +4,11 @@ namespace Dotmailer;
 
 use Symfony\Component\Yaml\Parser;
 
-class Config
+class Config extends AbstractConfig
 {
-    protected $config;
-
     public function __construct($config_file)
     {
         $parser = new Parser();
         $this->config = $parser->parse(file_get_contents($config_file));
-    }
-
-    public function __get($key)
-    {
-        if (isset($this->config[$key])) {
-            return $this->config[$key];
-        } else {
-            return null;
-        }
     }
 }
